+++ src/Markdig/Extensions/AutoLinks/AutoLinkOptions.cs
@@ -13,5 +13,10 @@ namespace Markdig.Extensions.AutoLinks
         /// Should the link open in a new window when clicked (false by default)
         /// </summary>
         public bool OpenInNewWindow { get; set; }
+
+        /// <summary>
+        /// Should a www link be prefixed with https:// instead of http:// (false by default)
+        /// </summary>
+        public bool UseHttpsPrefixForWWWLinks { get; set; }
     }
 }